erro de propriedade no string
