+++ src/models/aeroporto_models.js
@@ -1,18 +1,18 @@
-const db = require("./db.js")
+const db = require("../db.js")
 
 const Schema = db.Schema;
 
 const aeroportoSchema = new Schema({
     nome : {    
-        type: Schema.types.String,
+        type: Schema.types.String(),
         required: true
     },
     endereco: {
-        type: Number
+        type: Number()
     },
 })
 
-const aeroporto = db.model("index", aeroportoSchema)
+const aeroporto = db.model("aeroporto", aeroportoSchema)
 
 
 module.exports = aeroporto;
